r518: merge in the SAMBA_2_2 branch from cvs to brnaches/SAMBA_2_2 this is maybe not complete yet, please wait until I create the DAY_ZERRO tag
metze

+++ source/lib/sendfile.c
@@ -38,7 +38,7 @@ ssize_t sys_sendfile(int tofd, int fromfd, const DATA_BLOB *header, SMB_OFF_T of
 {
 	size_t total=0;
 	ssize_t ret;
-	size_t hdr_len = 0;
+	ssize_t hdr_len = 0;
 
 	/*
 	 * Send the header first.
@@ -161,7 +161,7 @@ ssize_t sys_sendfile(int tofd, int fromfd, const DATA_BLOB *header, SMB_OFF_T of
 
 		vec[0].sfv_fd = SFV_FD_SELF;
 		vec[0].sfv_flag = 0;
-		vec[0].sfv_off = (off_t)header->data;
+		vec[0].sfv_off = header->data;
 		vec[0].sfv_len = hdr_len = header->length;
 
 		vec[1].sfv_fd = fromfd;
